requirements.txt
tensorflow>=2.20.0
keras>=3.12.0
opencv-python-headless>=4.12.0
pandas>=2.3.0
numpy>=2.2.0
scikit-learn>=1.7.0
matplotlib>=3.10.0
seaborn>=0.13.0
pyyaml>=6.0.0
tqdm>=4.67.0
pillow>=12.0.0
